refactor: Consolidate init files into 1 & modernize
Instead of having many files for different paths with substantially the
same code, all the common code was combined into a single script. The
differences were added as switches based on an input argument to the
script.

Error handling for the arguments was included. Where certain commands
could result in errors printed to the terminal, the stderr output was
passed to /dev/null to keep the terminal free from worrisome errors that
mean nothing to the user.

The system bash calls were also modernized with the latest recommended
practice as of Py3.5. This means using subprocess.run wherever possible.
Part of the proper use of this is to the returncode for success or
failure, and to explicitly check (or not check) for errors in the
execution of the command. In this case, all are false.
This is referenced in: https://github.com/PyCQA/pylint/issues/2848

Also included are pauses in script execution to allow time for the flask
webserver to fully load before loading the first html page. This removes
the need for the user to manually reload the webpage (this comes across
as a broken user experience).

Finally, originally the many scripts all had the
`snap install network-manager`. This was unnecessary and undesirable for
a couple reasons:
1. Only the 2nd PC in the YetiCold mode needs to have the network
disabled. No where else is it needed.
2. After network-manager is installed, it disables the network (even
though everything look ok). This is also very bad for user experience
because it is not obvious that the network is down.

In the future, this practice should be reconsidered for a different
approach, even if manual with specific directions and warnings.