Clean up code & add docstring
'rm ~/testblockchain' wasn't needed as that folder is never created.

Use os.path.join for contructing folder and file paths in an agnostic
way.

Convert the subprocess.call functions into
subprocess.run functions

Replace ; in bash command with && so it only runs if the previous was
successful

Change `cd` to `cd ~` for clarity and explicitness