refactor: Rework network disable process
In Ubuntu 18 and 20, network-manager is already installed.
There is no need to install it with SNAP and doing so silently disables
the network and leaves the user wondering what's wrong. They would need
to reboot in order for the network to work so the installation can
continue.

Also, the network disable commands were consolidated into the
forgetnetworks.py script for consistency and maintainance.