WIP - Don't merge.  Read comments.
If I'm reading the code right, it looks like all 3 levels of wallets direct to this page? But the description doesn't apply well to all three levels (Level 1 is a single sig, level2 and level3 is multi sig).  The changes here need to apply for the Level 1 wallet, but a new page should be created for Level 2 and Level 3 wallets to direct to (and should be the old page describing requiring 3 privkeys).  The spelling change should apply to both pages.